Move segyio/_line.py -> segyio/line.py
