+++ python/segyio/line.py
